refactored entities and dtos to belong to model package , added a mapper for converting from and to a DTO
