+++ src/main/java/com/example/model/entity/Item.java
@@ -1,4 +1,4 @@
-package com.example.entity;
+package com.example.model.entity;
 
 import io.micronaut.serde.annotation.Serdeable;
 import jakarta.persistence.*;
