+++ src/main/java/com/example/model/entity/OrderItemId.java
@@ -1,4 +1,4 @@
-package com.example.entity;
+package com.example.model.entity;
 
 
 import lombok.AllArgsConstructor;
